britecore/bcinsurer/bcinsurer-requirements.txt
codegen==1.0
Django==1.11.10
djangorestframework==3.7.7
patterns==0.3
psycopg2==2.7.1
pytz==2017.3

